add test

+++ src/models/index.js
@@ -4,14 +4,17 @@ const dbConfig = require('../config/dbConfig');
 const sequelize = new Sequelize(dbConfig.DB, dbConfig.USER, dbConfig.PASSWORD, {
   host: dbConfig.HOST,
   dialect: 'mysql',
-  operatorsAliases: false
+  operatorsAliases: 0,
+  logging: false
 });
-
-sequelize.authenticate().then(
-  console.log(`Database connect to ${dbConfig.DB} on ${dbConfig.HOST}`)
-).catch((err) => {
-  console.error(`Error: ${err.message}`);
-});
+const authenticateDB = async () => {
+  try {
+    await sequelize.authenticate();
+  } catch (error) {
+    console.error('Unable to connect to the database:', error);
+  }
+};
+authenticateDB();
 
 const db = {};
 
@@ -21,17 +24,28 @@ db.sequelize = sequelize;
 db.users = require('./userModel')(sequelize, DataTypes);
 db.questions = require('./questionModel')(sequelize, DataTypes);
 db.answers = require('./answerModel')(sequelize, DataTypes);
+db.votes = require('./voteModel')(sequelize, DataTypes);
 db.comments = require('./commentModel')(sequelize, DataTypes);
 
 // Associations
 db.users.hasMany(db.questions);
 db.users.hasMany(db.answers);
 db.users.hasMany(db.comments);
+db.questions.belongsTo(db.users, { as: 'author', foreignKey: 'userId' });
+db.answers.belongsTo(db.users, { as: 'author', foreignKey: 'userId' });
+db.comments.belongsTo(db.users);
 db.questions.hasMany(db.answers, { onDelete: 'cascade', allowNull: false });
 db.answers.hasMany(db.comments, { onDelete: 'cascade', allowNull: false });
+db.answers.hasMany(db.votes, { foreignKey: 'answerId' });
 
-db.sequelize.sync({ force: true }).then(() => {
-  console.log('yes re-sync done');
-});
+// db.sequelize.sync({ force: true, match: /_test$/ }).then(() => {
+//   console.log('yes re-sync done');
+// });
+
+if (process.env.NODE_ENV !== 'test') {
+  db.sequelize.sync({ force: true }).then(() => {
+    console.log('yes re-sync done');
+  });
+}
 
 module.exports = db;
